Refactor namespaces structure and clean code for users account logic
