+++ src/FoodMaster.WebSite/Commands/LoginUser/LoginCredentials.cs
@@ -1,12 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
+﻿using System.ComponentModel.DataAnnotations;
 using System.Security.Claims;
-using System.Threading.Tasks;
+
 using MediatR;
 
-namespace FoodMaster.WebSite.Models
+namespace FoodMaster.WebSite.Commands.LoginUser
 {
     public class LoginCredentials : IRequest<Claim[]>
     {
