+++ src/FoodMaster.WebSite/Commands/RegisterUser/RegisterUserRequestHandler.cs
@@ -6,46 +6,43 @@ using System.Threading.Tasks;
 
 using FoodMaster.WebSite.Abstraction.Services;
 using FoodMaster.WebSite.Domain;
-using FoodMaster.WebSite.Models;
 
 using MediatR;
 
-namespace FoodMaster.WebSite.Commands
+namespace FoodMaster.WebSite.Commands.RegisterUser
 {
-
-    public class CreateGuestCommandHandler : IRequestHandler<GuestCredentials, Claim[]>
+    public class RegisterUserRequestHandler : IRequestHandler<UserDetails, Claim[]>
     {
         private readonly IUsersService usersService;
 
-        public CreateGuestCommandHandler(IUsersService usersService)
+        public RegisterUserRequestHandler(IUsersService usersService)
         {
             this.usersService = usersService;
         }
 
-        public async Task<Claim[]> Handle(GuestCredentials request, CancellationToken cancellationToken)
+        public async Task<Claim[]> Handle(UserDetails request, CancellationToken cancellationToken)
         {
             var userId = Guid.NewGuid().ToString();
-            var assignedRole = Roles.Guest.ToString();
+            var assignedRole = Roles.User.ToString();
 
             var claims = new Claim[]
             {
                 new Claim(ClaimTypes.NameIdentifier, userId),
                 new Claim(ClaimTypes.Name, request.FullName),
-                new Claim(ClaimTypes.DateOfBirth, request.BirthDate.ToString()),
                 new Claim(ClaimTypes.Role, assignedRole)
             };
 
             var user = new User
             {
                 Id = userId,
-                UserName = userId,
+                UserName = request.UserName,
                 FullName = request.FullName,
                 BirthDate = request.BirthDate,
                 Claims = new List<Claim>(claims),
                 Role = assignedRole
             };
 
-            usersService.Create(user);
+            usersService.Create(user, request.Password);
 
             return await Task.FromResult(claims);
         }
